+++ dotnet/src/SmartStandards.Logging.Centralized.net48/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("2.11.2")]
-[assembly: AssemblyFileVersion("2.11.2")]
+[assembly: AssemblyVersion("2.11.3")]
+[assembly: AssemblyFileVersion("2.11.3")]